Add touch support for Android using new Input System
Due to the use of the new Input System, I could further refactor the
code to make it more modular reduce cross references.
For simplicity I decided to use a Singleton for the InputManager, even
though it is not strictly necessary, since I could manually inject
dependecies to the scripts that use it.

Tested it in Anderoid and it works fine, except camera is moving too
fast (need adjustments)